+++ pkg/frontend/querymiddleware/roundtrip.go
@@ -8,6 +8,7 @@ package querymiddleware
 import (
 	"context"
 	"flag"
+	"fmt"
 	"net/http"
 	"strconv"
 	"strings"
@@ -28,6 +29,9 @@ const (
 	day                    = 24 * time.Hour
 	queryRangePathSuffix   = "/query_range"
 	instantQueryPathSuffix = "/query"
+
+	cacheResultsFlagName      = "query-frontend.cache-results"
+	maxSeriesPerShardFlagName = "query-frontend.query-sharding-max-series-per-shard"
 )
 
 // Config for query_range middleware chain.
@@ -35,10 +39,11 @@ type Config struct {
 	SplitQueriesByInterval time.Duration `yaml:"split_queries_by_interval" category:"advanced"`
 	AlignQueriesWithStep   bool          `yaml:"align_queries_with_step"`
 	ResultsCacheConfig     `yaml:"results_cache"`
-	CacheResults           bool `yaml:"cache_results"`
-	MaxRetries             int  `yaml:"max_retries" category:"advanced"`
-	ShardedQueries         bool `yaml:"parallelize_shardable_queries"`
-	CacheUnalignedRequests bool `yaml:"cache_unaligned_requests" category:"advanced"`
+	CacheResults           bool   `yaml:"cache_results"`
+	MaxRetries             int    `yaml:"max_retries" category:"advanced"`
+	ShardedQueries         bool   `yaml:"parallelize_shardable_queries"`
+	CacheUnalignedRequests bool   `yaml:"cache_unaligned_requests" category:"advanced"`
+	MaxSeriesPerShard      uint64 `yaml:"query_sharding_max_series_per_shard" category:"experimental"`
 
 	// CacheSplitter allows to inject a CacheSplitter to use for generating cache keys.
 	// If nil, the querymiddleware package uses a ConstSplitter with SplitQueriesByInterval.
@@ -50,9 +55,10 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
 	f.IntVar(&cfg.MaxRetries, "query-frontend.max-retries-per-request", 5, "Maximum number of retries for a single request; beyond this, the downstream error is returned.")
 	f.DurationVar(&cfg.SplitQueriesByInterval, "query-frontend.split-queries-by-interval", 24*time.Hour, "Split range queries by an interval and execute in parallel. You should use a multiple of 24 hours to optimize querying blocks. 0 to disable it.")
 	f.BoolVar(&cfg.AlignQueriesWithStep, "query-frontend.align-queries-with-step", false, "Mutate incoming queries to align their start and end with their step.")
-	f.BoolVar(&cfg.CacheResults, "query-frontend.cache-results", false, "Cache query results.")
+	f.BoolVar(&cfg.CacheResults, cacheResultsFlagName, false, "Cache query results.")
 	f.BoolVar(&cfg.ShardedQueries, "query-frontend.parallelize-shardable-queries", false, "True to enable query sharding.")
 	f.BoolVar(&cfg.CacheUnalignedRequests, "query-frontend.cache-unaligned-requests", false, "Cache requests that are not step-aligned.")
+	f.Uint64Var(&cfg.MaxSeriesPerShard, maxSeriesPerShardFlagName, 0, "How many series a single sharded partial query should load at most. This is not a strict requirement guaranteed to be honoured by query sharding, but a hint given to the query sharding when the query execution is initially planned. 0 to disable cardinality-based hints.")
 	cfg.ResultsCacheConfig.RegisterFlags(f)
 }
 
@@ -64,6 +70,10 @@ func (cfg *Config) Validate() error {
 		}
 		if err := cfg.ResultsCacheConfig.Validate(); err != nil {
 			return errors.Wrap(err, "invalid ResultsCache config")
+		}
+	} else {
+		if cfg.MaxSeriesPerShard > 0 {
+			return fmt.Errorf("-%s may only be enabled in conjunction with -%s", maxSeriesPerShardFlagName, cacheResultsFlagName)
 		}
 	}
 	return nil
@@ -173,9 +183,9 @@ func newQueryTripperware(
 		queryRangeMiddleware = append(queryRangeMiddleware, newInstrumentMiddleware("step_align", metrics, log), newStepAlignMiddleware())
 	}
 
+	var c cache.Cache
 	// Inject the middleware to split requests by interval + results cache (if at least one of the two is enabled).
 	if cfg.SplitQueriesByInterval > 0 || cfg.CacheResults {
-		var c cache.Cache
 
 		// Init the cache client.
 		if cfg.CacheResults {
@@ -221,14 +231,32 @@ func newQueryTripperware(
 	)
 
 	if cfg.ShardedQueries {
+		// Inject the cardinality estimation middleware after time-based splitting and
+		// before query-sharding so that it can operate on the partial queries that are
+		// considered for sharding.
+		if cfg.MaxSeriesPerShard > 0 {
+			cardinalityEstimationMiddleware := newCardinalityEstimationMiddleware(c, log, registerer)
+			queryRangeMiddleware = append(
+				queryRangeMiddleware,
+				newInstrumentMiddleware("cardinality_estimation", metrics, log),
+				cardinalityEstimationMiddleware,
+			)
+			queryInstantMiddleware = append(
+				queryInstantMiddleware,
+				newInstrumentMiddleware("cardinality_estimation", metrics, log),
+				cardinalityEstimationMiddleware,
+			)
+		}
+
 		queryshardingMiddleware := newQueryShardingMiddleware(
 			log,
 			engine,
 			limits,
+			cfg.MaxSeriesPerShard,
 			registerer,
 		)
-		queryRangeMiddleware = append(
-			queryRangeMiddleware,
+
+		queryRangeMiddleware = append(queryRangeMiddleware,
 			newInstrumentMiddleware("querysharding", metrics, log),
 			queryshardingMiddleware,
 		)