+++ pkg/querier/blocks_store_queryable_test.go
@@ -480,67 +480,11 @@ func TestBlocksStoreQuerier_Select(t *testing.T) {
 					}}: {block1, block2},
 				},
 			},
-			limits:       &blocksStoreLimitsMock{maxChunksPerQuery: 1},
-			queryLimiter: noOpQueryLimiter,
-			expectedErr:  validation.LimitError(fmt.Sprintf(maxChunksPerQueryLimitMsgFormat, fmt.Sprintf("{__name__=%q}", metricName), 1)),
-		},
-		"max chunks per query limit hit while fetching chunks at first attempt - global limit": {
-			finderResult: bucketindex.Blocks{
-				{ID: block1},
-				{ID: block2},
-			},
-			storeSetResponses: []interface{}{
-				map[BlocksStoreClient][]ulid.ULID{
-					&storeGatewayClientMock{remoteAddr: "1.1.1.1", mockedSeriesResponses: []*storepb.SeriesResponse{
-						mockSeriesResponse(series1Label, minT, 1),
-						mockSeriesResponse(series1Label, minT+1, 2),
-						mockHintsResponse(block1, block2),
-					}}: {block1, block2},
-				},
-			},
 			limits:       &blocksStoreLimitsMock{},
 			queryLimiter: limiter.NewQueryLimiter(0, 0, 1),
 			expectedErr:  validation.LimitError(fmt.Sprintf(limiter.MaxChunksPerQueryLimitMsgFormat, 1)),
 		},
 		"max chunks per query limit hit while fetching chunks during subsequent attempts": {
-			finderResult: bucketindex.Blocks{
-				{ID: block1},
-				{ID: block2},
-				{ID: block3},
-				{ID: block4},
-			},
-			storeSetResponses: []interface{}{
-				// First attempt returns a client whose response does not include all expected blocks.
-				map[BlocksStoreClient][]ulid.ULID{
-					&storeGatewayClientMock{remoteAddr: "1.1.1.1", mockedSeriesResponses: []*storepb.SeriesResponse{
-						mockSeriesResponse(series1Label, minT, 1),
-						mockHintsResponse(block1),
-					}}: {block1, block3},
-					&storeGatewayClientMock{remoteAddr: "2.2.2.2", mockedSeriesResponses: []*storepb.SeriesResponse{
-						mockSeriesResponse(series2Label, minT, 2),
-						mockHintsResponse(block2),
-					}}: {block2, block4},
-				},
-				// Second attempt returns 1 missing block.
-				map[BlocksStoreClient][]ulid.ULID{
-					&storeGatewayClientMock{remoteAddr: "3.3.3.3", mockedSeriesResponses: []*storepb.SeriesResponse{
-						mockSeriesResponse(series1Label, minT+1, 2),
-						mockHintsResponse(block3),
-					}}: {block3, block4},
-				},
-				// Third attempt returns the last missing block.
-				map[BlocksStoreClient][]ulid.ULID{
-					&storeGatewayClientMock{remoteAddr: "4.4.4.4", mockedSeriesResponses: []*storepb.SeriesResponse{
-						mockSeriesResponse(series2Label, minT+1, 3),
-						mockHintsResponse(block4),
-					}}: {block4},
-				},
-			},
-			limits:       &blocksStoreLimitsMock{maxChunksPerQuery: 3},
-			queryLimiter: noOpQueryLimiter,
-			expectedErr:  validation.LimitError(fmt.Sprintf(maxChunksPerQueryLimitMsgFormat, fmt.Sprintf("{__name__=%q}", metricName), 3)),
-		},
-		"max chunks per query limit hit while fetching chunks during subsequent attempts - global": {
 			finderResult: bucketindex.Blocks{
 				{ID: block1},
 				{ID: block2},